+++ src/common/hook/Modal/useModal.tsx
@@ -1,22 +1,22 @@
-import { ReactElement, useState } from 'react';
+import { ReactElement, useCallback, useState } from 'react';
 
-import usePreventGoBack from '@/common/hooks/Modal/usePreventGoBack';
+import usePreventGoBack from '@/common/hook/usePreventGoBack';
 
 const useModal = () => {
   const [isOpen, setIsOpen] = useState(false);
   const [currentContent, setCurrentContent] = useState<ReactElement | undefined>(undefined);
 
   // 열기
-  const openModal = (content: ReactElement) => {
+  const openModal = useCallback((content: ReactElement) => {
     setCurrentContent(content);
     setIsOpen(true);
-  };
+  }, []);
 
   // 닫기
-  const closeModal = () => {
+  const closeModal = useCallback(() => {
     setIsOpen(false);
     setCurrentContent(undefined);
-  };
+  }, []);
 
   // 뒤로가기 이벤트 처리
   usePreventGoBack(isOpen, closeModal);
